Move cmd/pivit/yubikey/yubikey_identity.go -> pkg/pivit/yubikey/yubikey_identity.go
