+++ cmd/pivit/import.go
@@ -5,8 +5,8 @@ import (
 	"encoding/pem"
 	"os"
 
-	"github.com/cashapp/pivit/cmd/pivit/yubikey"
 	"github.com/cashapp/pivit/pkg/pivit/utils"
+	"github.com/cashapp/pivit/pkg/pivit/yubikey"
 	"github.com/go-piv/piv-go/piv"
 	"github.com/pkg/errors"
 )
