+++ cmd/pivit/reset.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"math/big"
 
-	"github.com/cashapp/pivit/cmd/pivit/yubikey"
 	"github.com/cashapp/pivit/pkg/pivit/utils"
+	"github.com/cashapp/pivit/pkg/pivit/yubikey"
 	"github.com/go-piv/piv-go/piv"
 	"github.com/pkg/errors"
 )
