+++ cmd/pivit/sign.go
@@ -8,9 +8,9 @@ import (
 	"os"
 	"strings"
 
-	"github.com/cashapp/pivit/cmd/pivit/yubikey"
 	"github.com/cashapp/pivit/pkg/pivit/status"
 	"github.com/cashapp/pivit/pkg/pivit/utils"
+	"github.com/cashapp/pivit/pkg/pivit/yubikey"
 	cms "github.com/github/smimesign/ietf-cms"
 	"github.com/pkg/errors"
 )
